Move entity specs to model and remove DerivesFrom ConstraintType
